GUACAMOLE-210: Move classes to reasonable packages.
